OMAP3+: xSI: Add FIFO flush after R/W cancel
When hsi_read_cancel is called, it simply cancels the DMA or acknowledge the RX interrupt.
But data in the FIFO remains present, which can cause next hsi_read to contain up to 8
garbage (old) frames (as FIFO size is 8 frames) in the beginning of the reception.

Same for TX side.

Change-Id: I946d05120815def27fe401677e24c37ca8a6532e
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>