ASoC: OMAP4 ABE: Allow an alternate ABE firmware to be used.
On commit 52c46046 the ABE firmware was updated with a version that
included several improvements, but also disabled the digital
microphone (DMIC) feature.  In order to support DMIC on devices like
the Blaze, an alternate firmware (and matching code) must be
configured.

This commit allows an alternate firmware using
CONFIG_SND_OMAP4_ABE_USE_ALT_FW.

Change-Id: I2fd4aa8d03be802722dca973a71435b9965e80ea
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>