Input: omap4-keypad - cleanup keypad driver
Long key press is not needed on android since it
is simulated on user space. Extra interrupts may
leave userspace confused.

Use bitfiled instead of hardcoded values to set
KBD_CTRL, and change PVR divisor to 0x6, which
would allow a debouncing time of ~10ms.

Set mpu irq to level instead of edge, since if mpu
is in low power a edge detection may be lost if
the event is a wkup event.

Change-Id: I2b96abdfa573614a5a669b13eed8c60baf2b07af
Signed-off-by: Andrii Danylov <x0158318@ti.com>
Signed-off-by: Axel Haslam <axelhaslam@ti.com>