OMAP4: Avoid clearing i2c2 irq on A9 when HWSPINLOCK is enabled
This patch is related to this commit:
7dd92cb9a8b9c6bb0f4b03afd26ea50ec90cfb73

On Blaze and Blaze Tablet, the i2c-2 bus is shared by A9 and Ducati.
We want A9 to be able to access the i2c-2 bus when it needs it.
This patch will prevent the i2c-2 irq from being cleared on A9 when
Ducati requests the i2c-2 resource, since hwspinlock is being used to
arbitrate the use of the i2c-2 bus.

Change-Id: Iddb667f77e17d9799a9caa62d9f2a0a0043f1768
Signed-off-by: Christina Warren <cawarren@ti.com>