usb: otg: Handle spurious ID GND interrupt
On Blaze tablet boards it is observed that sometimes
in trying to plugin a micro-A type plug it is causing
consecutive ID GND interrupts which imbalances the PM
get_syc() counters. Because of the design of the
front panel/otg port on tablet boards, it is difficult
to attach/detach a cable smoothly.

To avoid this scenario, we ignore the ID GND interrupt
if one has already been served without a detach seen.

Change-Id: I1b034a6692f4e381d1a89b729cc942b763265396
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>