OMAP4: HDMI: Add suspend handler for LCD panel
Add a suspend handler for LCD panel.
HDMI suspend/ resume path is broken due to lack of a suspend handler
for LCD panel. When a HPD signal is received
during resume it sends a device switch setting to user space causing
set_mode call from user space for reconfiguring the HDMI panel.
this enables the LCD panel but when the system goes to suspend again
it fails as the LCD doesn't have a suspend handler. LCD is currently
being disabled / enabled from DSSCOMP as part of Early suspend/Late Resume
But if the LCD is enabled after early suspend is completed there is no
way to disable it. This patch will fix this issue.
Remove the device switch setting for HDMI panel as this wiil cause
the user space to wake up for handling the switch setting and result
in the suspend path being broken.

Change-Id: Ia6c003d096fdce96935158fb17e9e079e0b8705a
Signed-off-by: Srinivas Pulukuru <srinivas.pulukuru@ti.com>