USB: OTG : Wake lock holding decision based on usb connect/disconnect
Descpription: Currently the wakelock is held based on usb_cinlimit_mA.
Expected behaviour is that when the cable is disconnected usb_cinlimit_mA
should be zero. But some times when the USB cable is disconnected usb_cinlimit_mA
is still not zero. An additional condition is added to not hold the wake_lock in such
scenerio.

Change-Id: Id7815018354b35ff87224481047d4ee0365d709f
Signed-off-by: Shivasharan Nagalikar <shivasharan.nagalikar@ti.com>