OMAP4: USB: Update EHCI TLL errata i640
The workaround for EHCI TLL i640 errata has a critical timing window between
clearing the FPR bit in the PORTSC register and cutting the TLL clock. The
register writes to clear the FPR bit and cut the TLL clock need to occur in
back to back instructions otherwise we could still experience a disconnect
as described in the original errata. The ehci_writel function will cause a
L2 cache flush and so when compile takes multiple instructions. Instead of
using ehci_writel use __raw_readl.

Change-Id: Iaacf6987e37a1096b9d1c59fb3e71eb0f1c7aac3
Signed-off-by: Jon Hunter <jon-hunter@ti.com>