OMAP4:To Fix the Slow Resume during first suspend
The below patch gave a workaround to avoid Long Key Press to resume
from suspend
http://review.omapzoom.org/#change,17913
But after reboot if we go to suspend without pressing the power key
i.e. waiting for timeout, then its not resuming normally.
So with this patch, if its first time we provide press/release
event to the upper layer.

Change-Id: If4d44adcc3e95bdaf8a5eec16b3a7561ae8a7b2a
Signed-off-by: Siva Pothireddy <sivakumar.pothireddy@ti.com>