MFD: twl6040-codec: Reinitialize completion in automatic power-up
TWL6040 automatic power-up sequence uses the completion interface.
The completion is marked when CODEC driver receives the READYINT
which indicates automatic sequence has finished.

When the system is highly loaded it has been seen that the CODEC
threaded interrupt handler (and complete() call) runs after timeout
has expired. If the retry condition is met, then the call to
wait_for_completion_timeout() will immediately return because of
the previous completion. Therefore it's required to reinitialize
the completion structure.

Change-Id: I6bb0c3149a1e2d03f67063cd1664147aff044f89
Signed-off-by: Sivakumar Pothireddy <sivakumar.pothireddy@ti.com>
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>