POWER: twl6030_bci_battery fix miscompile in start
For some reason some versions of the CSL compiler mis-compile the code
in twl6030_start_usb_charger. Re-order the code to make compiler
happier. In original the compiler ends up generating code which sends
a zero to the write function in all cases. Assigning to a variable
and using the new form the correct value gets written. The new version
is also easier to read. This issue was reported by a customer
and seen locally.

The conditional was also wrong, the correct conditional is use_power_path
not just the fact this is TWL6032.

Change-Id: I175335eec8df4f9fb9660ce0430441c5737d8e07
Signed-off-by: Graeme Gregory <gg@slimlogic.co.uk>