OMAP4: L2C: Restore L2 POR on GP devices
Current code restores POR on HS devices only. Add support for restoring POR on
GP devices.

Failure to restore POR severely affects memory bandwidth on the device.

This was found during investigating a memory throughput issue on GP devices
using the 'stream' benchmark as reported by Rocky Rhodes. On GP devices the
memory throuput was slashed by almost half after the first OSWR to ON
transition was done. Restoring POR fixes that issue.

The original patch for this was done by Volodymyr Riazantsev on
p-android-omap-2.6.35 (Commit 34d50a29) which seems to have been lost in the
newer kernel versions. Hopefully this one is not lost again :-)

[vikram pandita]
	- fixed bug with decimal vs hex in assembly
	- made code cleaner by passing 0x443022 to mean 4430 ES2.2
	- fix por setting on 4430 on ES2.2 and above silicon only

Change-Id: I3b607d2436e8fc5c5c1719a726b8fa88e339625c
Cc: Santosh Shilimkar <santosh.shilimkar@ti.com>
Cc: Richard Woodruff <r-woodruff2@ti.com>
Signed-off-by: Vimarsh Zutshi <vimarsh.zutshi@ti.com>
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>