omap: add dsp carveout for remoteproc functionality
remoteproc uses DDR carveout memories for the functionality
of different rproc instances. New interfaces have been added
to perform this carveout memory during board initialization
for DSP. This patch reserves two types of carveout memory -
one whose address is dynamic, and another that leverages a
fixed static range. The dynamic pool will be used for
allocating memory for firmware sections that have no physical
memory allocated. The static pool will be used to validate the
firmware sections with fixed physical addresses.

This patch also assigns addresses for the Virtio vring buffers
for DSP. This is previously not handled, and the buffers are
currently assigned from the static pool associated with the
dsp. The rpmsg bus initialization sequence is enhanced to
process all the vprocs even if an earlier one in the list
fails due to memory requirements.

Change-Id: I0dec512b63904d3ffe0867ce00150b6c21a80805
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>