remoteproc: ignore static pool check for dsp IO buffers
The dsp remoteproc is using the same pool of 1D buffer
region as the ducati. This cannot be configured in the
static pool of dsp as this memory region is co-located
only with ducati memory, but not dsp. So, ignore the
valid memory check to proceed with the loading & booting
for dsp.

Change-Id: Ia92a2b7329cd0b9431afad4dde8c83396d861ba2
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Jesse Villarreal <jesse.villarreal@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>