MFD: TWL6030: Change return result of the GPADC interrupt call back function
Previously was not possible to read the corrected and converted values
of the gpadc conversion in the call back function. Now pointer
to the twl6030_gpadc_request returned to the call back function
and we can read all adc channel values: raw without correction
and convertation, corrected and converted, raw values etc.

Change-Id: I4a3699952ddf9cacc4cca5e205b93e28502f291d
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>