TI thermal framework: omap die governor: introduce enter zone helper function
This code changes the OMAP DIE governor to use a helper function to enter
a thermal zone, instead of using a function for each zone. Idea is to
reduce the amount of duplicated code.

This code also introduces a data structure to represent a thermal zone.
The enter zone helper function is based on this data structure.

The logic alert zone logic computation, which depends on if we
are rising or lowering in temperature, is now moved to the zone
matching function. This place is much more convenient, then inside
the enter zone helper.

Change-Id: Ibcb9228556c793104d2ec6f8df092a1c7ea99966
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>