TI thermal framework: fix initialization sequence and introduce "pcb" domain
This patch creates a new thermal domain, named "pcb". Also it teaches
the omap die governor how to be aware of a separated domain.

The pcb sensor will not be part of the "cpu" domain, but instead, it
is now part of the "pcb" domain. This will change the system so that,
instead of wrongly attempt to register two sensors on the same domain,
they will be registered on two different domains, but the governor
is aware of both domains.

By doing so, it is not necessary to keep track of sensors being
registered on the same domain, and deciding by looking up the
sensor names.

Now that governors can fetch the temperature from a different domain,
there is no need to keep feeding the framework with temperature reports
by means of running workqueues. This workaround was necessary
because there was an initialization issue while registering several
sensors in the same domain.

Now, because there are multiple domains, there is no need to keep
track on sensors being registered and their reports. For now,
these workqueues are not needed.

Change-Id: I6e2764fc1d49fbe97953c2a9a5192f8585c91aab
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>