ARM: OMAP: Voltage: PMIC data: rename vp_vddmin and vp_vddmax fields
We used to use PMIC data defined min and max voltages which were a
balance between what the SoC supports and what PMIC supports.
As a first step to ensure scalable solution, we split the meaning
to now indicate PMIC's max and min voltages indicated by min_volt
and max_volt.

All voltage values are retained and will be modified in next
series of patches.

Change-Id: I475c6094bb6efd0142431c14cd94a48e2a8ca199
[oleksandr.kozaruk@ti.com, nm@ti.com: port to Android kernel]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Oleksandr Kozaruk <oleksandr.kozaruk@ti.com>
Signed-off-by: Tero Kristo <t-kristo@ti.com>