omap: remoteproc: set the load boot address in activate function
This is needed because the register holding the boot address may
lose context if a deep c-state is reached (either during Device
OFF or during suspend) by the power domain where the register
resides.

Save the extracted bootaddr and then populate the boot address
register before taking the remoteproc out of reset in the activate
function to cover both rproc_start and resume paths.

Change-Id: I0ab3a70268877bcf55d053c9144e3fb834b58939
Signed-off-by: Miguel Vadillo <vadillo@ti.com>