OMAP: TILER: Optimize placement for NV12 buffers
This patch changes the behaviour when allocating 2D buffers.  If
the required alignment is less than 4KiB and a security token is set,
the allocation will be flagged specifically for this format.  All
subsequent buffers of the same format and size will be placed inside
this area until the area is exhausted.

For NV12 allocations, this means that all Y allocations will share
one area, and the UV allocations will share another.

Change-Id: I13d29257c0cbab627c2938d4623557918448ee86
Signed-off-by: Andy Gross <andy.gross@ti.com>