DRIVERS: MISC: change the DC cpufreq control to TF control
Duty cycle uses sys fs to control cpufreq. It is not safe.
The user(overclocker) can burn the device using simple script that
changes max_frequency. The thermal framework has own mechanism for
this and the patch changes the duty cycle mechanism to thermal
framework variant of cpufreq control.

Change-Id: Ifc0f8b81addb7524eb67d116e4780978739a67f3
Signed-off-by: Eugen Mandrenko <ievgen.mandrenko@ti.com>

Conflicts:
	arch/arm/mach-omap2/omap2plus-cpufreq.c