OMAP4: Regulator: Perform additional check for some LDOs
For TWL6032 revision < ES1.1 with EEPROM revision < rev56.0
LDO6 and LDOLN must be always ON because of the hardware
bug in the TWL6032. For TWL6032 revision >= ES1.1 with
EEPROM revision >= rev56.0 those LDOs can be off in sleep-mode.
If LDO6 or LDOLN is always on then SYSEN must be always on.
This patch performs additional checks.

Change-Id: I81e8e1a9dfdb628849277f4c3b4363bd0146cac4
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>