remoteproc: get suspend address from resource table
The suspend address is used as a flag to check if the remoteproc
has actually suspended. The address could vary according to the
memory map in the remoteproc. Changing the memory map in the
remote processor changes the value of the suspend address.
Passing the address in the resource table information avoids
hardcoding the value in the platform data of the remoteproc.
The value is instead assigned while parsing the resource table.

Change-Id: I698b81d61d391af85bc1556d2abc4c407c1db018
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>