USB: gadget: f_mtp: fix send ZLP packet algorithm.
If size of the data that sent from MTP-device to computer is a multiple of
wMaxPacketSize (usb endpoint) then at the end of transfer device must
send ZLP packet. Previously in f_mtp module a variable 'maxpacket' that
used to send ZLP packet was always equal to 512. For USB2.0 wMaxPacketSize
equals 512 but for USB1.1 wMaxPacketSize can be 64, 32, 16 or 8. Now variable
'maxpacket' contains real value of wMaxPacketSize in all USB modes and ZLP
packet is sent correctly for USB1.1.

Change-Id: I75606c954119dea2a76d955849a47208df0d2db0
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>