twl6040: errata: PDMUP State In Sleep/Deep Sleep Mode
When AUDPWRON is set to 0 at any time in the IC operation to
put it into sleep or deep sleep mode, there is a finite probability
that the value of PDMUP could be at logic 1 (VDDVIO). In this state,
the resistive pulldown on OMAP on this pin results in a current drain.

Adding new impacted revisions (ES1.3, ES2.0, ES2.2) to errata workaround.

Change-Id: I049fe85eb2e1e9d04c72a9820f3adcc25db35646
Signed-off-by: Vadym Shovkoplias <shovkoplias.vadym@ti.com>

Conflicts:
	arch/arm/mach-omap2/board-44xx-power.c