Revert "omap4: sdp/panda/tablet: reserve memory for dsp"
This reverts commit e699cc2c4cc3f369108d83370172138294f9840f.

DSP currently requires 4MB of memory to be set-aside to meet its
firmware requirements and can use either a fixed carveout or
a dynamically assigned carveout. The above commit uses static
carveouts for DSP in various board files. DSP does not have a
mandatory requirement to use this fixed carveout, and has been
switched over to use a dynamically allocated carveout address.
The above commit is therefore being reverted to free up unused
carveout memory.

This change simultaneously fixes a build error when the DSP
remoteproc is included in the build configuration. The build
was by the following commit:
	"OMAP4: ion: remove hardcoded carveout sizes"

Conflicts:

	arch/arm/mach-omap2/include/mach/omap4_ion.h

Change-Id: Iee9be7dbc19dd580fabce6ae966cf86d1f54b5cb
Signed-off-by: Jesse Villarreal <jesse.villarreal@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>

Conflicts:
	arch/arm/mach-omap2/board-4430sdp.c
	arch/arm/mach-omap2/board-44xx-tablet.c
	arch/arm/mach-omap2/board-omap4panda.c
	arch/arm/mach-omap2/include/mach/omap4_ion.h