remoteproc: config: enable dsp by default
The DSP processor is currently not enabled by default in the
menuconfig. This is being turned on by default to make sure
DSP is activated, and usable in the system.

A DSP binary, tesla-dsp.bin is expected to be present in
the firmware folder for a successful load of the DSP remote
processor.

Change-Id: Ibd305a5a22e98cd09e59e037e428c49a4242abd2
Signed-off-by: Suman Anna <s-anna@ti.com>