gcx: fix race conditions on suspend.
There was a race between the suspend thread waiting for idle and an
app thread attempting to push more into the gcqueue, which could
cause the suspend to block until the max retry count was reached.
Also, it was possible for the gccmdthread to notice gc320 is idle
while attempting to suspend and simultanesouly gc320 completed
processing some cmd buffers that require proper interrupt-aware
cleanup.

Change-Id: Ia7fdb29c5e2936d07b4c7ce58638b57c1f10c955
Signed-off-by: Craig Stout <craig.stout@ti.com>