drivers: iio: inv mpu: don't power up mpu unnecessarily
Previously it would always power it back on when resuming from suspend.
Now on resume, it restores the state it had on suspend.
This reduces the time spent resuming. The total resume time was 250ms,
130ms was in inv mpu power up.

Even when the mpu is on to handle screen rotation, it gets turned off
when the device screen is turned off before it goes to sleep.
So this change is sufficient to remove the 130ms.
Use cases where some other service keeps on running and needs the mpu
will add back the 130ms. If really needed, then we can do something
fancier with tasks.

Bug: 7139924, 7309244
Signed-off-by: JP Abgrall <jpa@google.com>