omap: support TWL6032 PMIC and OMAP4460 combination in omap_twl_init()
This patch uses the PMIC type to determine if a fixup of the
omap_twl_map struct is needed to support a different CPU and PMIC
combination. Previously only OMAP4460+TWL6030+TPS62361 were supported,
now OMAP4460+TWL6032 is supported in addition as well.

Detected at runtime to be backwards compatible.

Change-Id: Ib51431abc02f87ed48398ddd36bee9b2fcd53e53
Signed-off-by: Hans-Christian Egtvedt <egtvedt@samfundet.no>